+++ Validation-Manager-Web/src/main/webapp/WEB-INF/sun-jaxws.xml
@@ -1,4 +1,3 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <endpoints version="2.0" xmlns="http://java.sun.com/xml/ns/jax-ws/ri/runtime">
-  <endpoint implementation="net.sourceforge.javydreamercsw.validation.manager.web.service.VMWebService" name="VMWebService" url-pattern="/VMWebService"/>
 </endpoints>
